+++ static/js/bundles/portico.js
@@ -1,4 +1,4 @@
 import "./common.js";
-import "../translations.js";
+import "../i18n.js";
 import "../portico/header.js";
 import "../../styles/portico/portico-styles.scss";
